Update rubric for instructor training applications
The rubric was embedded in the application form, which was plain text. Converted to markdown to make easier to read and extracted rubric items to top of document. 

Added note that, although this version of the application was approved by the Trainers group, it has not yet been implemented (as of 5/15/17).