Add video as alternative to GitHub
Creates an alternative option for workshop website setup, so that those who do not have a GitHub account and do not currently wish to set one up can watch the instructional video on how to complete the setup process. This will help them to develop a mental model without demanding that they interact with GitHub immediately.